UTF-8 fixes for localize()
When looking up locale information using POSIX::localeconv(), the
returned data honors the chraracter encoding of the current locale.
This is a big problem if, for example, the currency_symbol is a
non-ascii character.  The correct thing to do is to decode the data from
localeconv() according to the current locale's character encoding.  This
can be a complicated thing to figure out, but forutnately Encode::Locale
does this for us.  As a result, Encode::Locale, and Encode are now
required modules.  In addition, Test::More::UTF8 is required so that
test output sets the utf8 pragma on the test output filehandles.

The localize tests have been clenaed up and added back in at this time.