rework BigInt bcmp() fix
bcmp() now just checks how many args it was called with.  If called with
three args, it is assumed to be called as a class method with two args.
If not, then its assumed bcmp() was called as an object method.

This makes bcmp() work as documented in Math::BigFloat without breaking
<=> overloading in both old and new Math::BigInt versions.